requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-openai
langchain-community
ddgs